requirements.txt
clipspy==1.0.2
tk==0.1.0
